git-svn-id: http://crux-framework.googlecode.com/svn@3680 a5d2bbaa-053c-11de-b17c-0f1ef23b492c

+++ branches/4.1.0/Crux/src/ui/widgets/org/cruxframework/crux/widgets/client/maskedtextbox/IntFilterFormatter.java
@@ -35,7 +35,7 @@ public class IntFilterFormatter extends FilteredTextBoxBaseFormatter
 	{
 		if (input instanceof Integer)
 		{
-			return Long.toString((Long) input);
+			return Integer.toString((Integer) input);
 		}
 		else
 		{
